+++ src/core/QueryManager.ts
@@ -79,7 +79,6 @@ import {
 } from '../data/readFromStore';
 
 import {
-  MutationBehavior,
   MutationQueryReducersMap,
   MutationQueryReducer,
 } from '../data/mutationResults';
@@ -218,14 +217,12 @@ export class QueryManager {
   public mutate<T>({
     mutation,
     variables,
-    resultBehaviors = [],
     optimisticResponse,
     updateQueries: updateQueriesByName,
     refetchQueries = [],
   }: {
     mutation: DocumentNode,
     variables?: Object,
-    resultBehaviors?: MutationBehavior[],
     optimisticResponse?: Object,
     updateQueries?: MutationQueryReducersMap,
     refetchQueries?: string[],
@@ -260,7 +257,6 @@ export class QueryManager {
       operationName: getOperationName(mutation),
       mutationId,
       optimisticResponse,
-      resultBehaviors,
       extraReducers: this.getExtraReducers(),
       updateQueries,
     });
@@ -281,7 +277,6 @@ export class QueryManager {
             document: mutation,
             operationName: getOperationName(mutation),
             variables,
-            resultBehaviors,
             extraReducers: this.getExtraReducers(),
             updateQueries,
           });